Allow /usr/lib/networkd-dispatcher scripts with /etc overrides
Make --script-dir take a colon-separated lists of directories,
like $PATH, and implement lookup so that scripts in earlier
directories override those with the same name in later
directories.

This modifies scripts_in_path() to take the subdirectory as
a second argument, so it can join it with each path component,
rather than just with the last directory in the path.

Fixes #39